requirements.txt
aiofiles==24.1.0
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachelib==0.13.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
checksumdir==1.2.0
click==8.1.8
contourpy==1.3.0
cryptography==45.0.5
cycler==0.12.1
distlib==0.3.9
distro==1.9.0
duckdb==1.3.2
et_xmlfile==2.0.0
fastapi==0.115.12
filelock==3.17.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-RESTful==0.3.10
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
fonttools==4.54.1
fsspec==2025.5.1
greenlet==3.2.3
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.10.0
joblib==1.5.1
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
msgspec==0.19.0
networkx==3.5
numpy==2.1.2
openai==1.86.0
openpyxl==3.1.5
packaging==24.1
pandas==2.3.0
patsy==1.0.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.0.0
platformdirs==4.3.6
playwright==1.52.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyee==13.0.0
pyparsing==3.2.0
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==4.1.0
setuptools==78.1.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.37
starlette==0.46.2
statsmodels==0.14.5
sympy==1.14.0
tabula-py==2.10.0
tabulate==0.9.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.4.0
uvicorn==0.34.3
virtualenv==20.29.1
Werkzeug==3.1.3
